Optimized source code and other small changes
In this version I created a new class called RootUtils, it has a method
to check if the root access is present.
I also moved the window options from the main to the frame file to clean
and optimize the code.
The donation button has been removed, and I changed some hard-coded
items into constante at the beginnig of the frame file; the window is
now resizable.

To do:
    - improve GUI, with a more standard look;
    - better management of the window's close button;
    - maybe the integration with a modified version of acpi_call.