Version 2.0: project redesign, better permission management
The Java program has been redesigned to be more object oriented, stable and with a better look. It no longer needs to be run as root because pkexec is now used in the scripts.
Now before making a call the program checks if the GPU driver is unloaded, in fact it blacklists the driver after the installation of acpi_call: supports radeon, amdgpu and nouveau.

Detailed changelog for every file:
- noGPUDriver.* :
Required for blacklisting radeon, amdgpu or nouveau driver.

- AboutFrame.java:
new JFrame to dispaly information about the Author and the program.

- LogFrame.java:
previously was inside the main frame, now it is in a separate class and is
used to display the outupt of the scripts.

- acpi_call_GUI_systemdFrame.java:
renamed to ACGSFrame;
now manages only the graphic components of the main window, it calls ACGSCore to perform operations to the system;
Added a Menu to main window;
new listener for mouse input.

- ACGSCore:
Better web browser compatibility, now it tries to open the default one;
More checks for installation and modules load/unload;
Better exceptions management.